use a provider for redux-async-connect
