+++ .eslintrc
@@ -33,13 +33,14 @@
     }
   },
   "globals": {
-    "__DEVELOPMENT__": true,
     "__CLIENT__": true,
-    "__SERVER__": true,
-    "__LOGGER__": false,
-    "__DISABLE_SSR__": true,
+    "__DATA_LOADER__": '',
+    "__DEVELOPMENT__": true,
     "__DEVTOOLS__": true,
     "__DEVTOOLS_IS_VISIBLE__": true,
+    "__DISABLE_SSR__": true,
+    "__LOGGER__": false,
+    "__SERVER__": true,
     "webpackIsomorphicTools": true
   }
 }
